dev-requirements.txt
autoflake
black
flake8
isort
pytest
pytest-cov
pytest-asyncio
requests
twine
python-dotenv
git+https://github.com/PyCQA/doc8
